Re-enable unit tests for utility functions and ingest.
Both of these tests were already working, and probably should have been
enabled earlier; a comment in the SConscript suggests someone wanted
to just disable all pex_harness-related tests, but got these as well.

The ingest test does require access to a remote server, but I've verified
that it skips the test and issues a warning when the server cannot be
contacted, so I think it's safe.