indexExposure: move operational parameter out of Config
indexExposure: get database name from butler

The spatial database will live under the control of the butler.

This removes the need for the user to specify a database name,
which makes it useful within another Task.

indexExposure: prepare for use within ProcessCcdTask

We want to run IndexExposureTask from within ProcessCcdTask so that
every CCD is indexed as it is processed, saving a separate I/O stage.
In order to do so, we need to move the operational parameter 'defer_writes'
out of the IndexExposureConfig, and put the database under the butler.

In order to preserve the IndexExposureTask as a standalone CmdLineTask,
moved the 'defer_writes' option into the ArgumentParser as
'--defer-writes'.

Unfortunately, at the moment, the butler doesn't handle databases, so
we will grab the database by filename and operate on it ourselves.